+++ src/main/java/com/kakaoasset/portfolio/controller/NewsStockController.java
@@ -1,6 +1,6 @@
-package com.kakaoasset.portfolio.elasticsearch.elasticAPI.controller;
+package com.kakaoasset.portfolio.controller;
 
-import com.kakaoasset.portfolio.elasticsearch.elasticAPI.service.NewsStockService;
+import com.kakaoasset.portfolio.service.NewsStockService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
